Evento cobra adicionado, bugs relacionados ao combate consertados, agora atirar gasta munição

+++ src/game/itens/food/Food.java
@@ -7,7 +7,7 @@ public abstract class Food extends Item {
     private int hungerPoints;  // pontos de fome restaurados
     private int expiration;  // dias ate estragar
 
-    public Food(String name, float weight, int durability, int hungerPoints, int expiration) {
+    public Food(String name, double weight, int durability, int hungerPoints, int expiration) {
         super("food", name, weight, durability, false);
         this.hungerPoints = hungerPoints;
         this.expiration = expiration;
